requirements.txt
# Core
anthropic>=0.30.0
python-dotenv>=1.0.0

# X / Twitter scraping
twikit>=2.1.0

# Web research
duckduckgo-search>=6.2.0

# Telegram notifications
python-telegram-bot>=21.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Scheduler
apscheduler>=3.10.4

# Async HTTP (used internally by twikit)
aiohttp>=3.9.0
